% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{addSpaceIfDuplicate}
\alias{addSpaceIfDuplicate}
\title{Add Space to Avoid Duplicate Elements}
\usage{
addSpaceIfDuplicate(x)
}
\arguments{
\item{x}{A \code{character vector}}
}
\description{
Add space to every element if there are duplicates in the vector.
}
\examples{
##no duplicate
x1 = c('simts2', 'simts1', 'simts3')
addSpaceIfDuplicate(x1)

##duplicate
x2 = c('simts3', 'simts4', 'simts3', 'simts4', 'simts5', 'simts6','simts3')
addSpaceIfDuplicate(x2)
}
\keyword{internal}
