% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{AR1}
\alias{AR1}
\title{Simulation of an Autoregressive Process of Order 1}
\usage{
AR1(phi = NULL, sigma2 = 1)
}
\arguments{
\item{phi}{A \code{double} value for the \eqn{\phi}{phi} of an AR1 process (see Note for details).}

\item{sigma2}{A \code{double} value for the variance, \eqn{\sigma ^2}{sigma^2}, of a WN process (see Note for details).}
}
\value{
An S3 object with called ts.model with the following structure:
\describe{
 \item{process.desc}{Used in summary: "AR1","SIGMA2"}
 \item{theta}{\eqn{\phi}{phi}, \eqn{\sigma^2}{sigma^2}}
 \item{plength}{Number of Parameters}
 \item{desc}{"AR1"}
 \item{obj.desc}{Depth of Parameters e.g. list(1,1)}
 \item{starting}{Guess Starting values? TRUE or FALSE (e.g. specified value)}
}
}
\description{
Simulation of an Autoregressive Process of Order 1
}
\note{
We consider the following model: \deqn{X_t = \phi X_{t-1} + \sigma \varepsilon_t,} where \eqn{\phi} is iid from a standard normal distribution.
}
\examples{
AR1()
AR1(phi=.32, sigma2=1.3)
}
\author{
James Balamuta
}
