% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{best_model}
\alias{best_model}
\title{Select the best model}
\usage{
best_model(x, ic = "aic")
}
\arguments{
\item{x}{An object from either \code{\link{select_arima}},
\code{\link{select_arma}}, \code{\link{select_ar}}, or \code{\link{select_ma}}.}

\item{ic}{Type of criterion to use in selecting the best model.}
}
\description{
Retrieves the best model
}
\examples{
sunspot.year \%>\%
 gts(start = 1700, freq = 1) \%>\%
 select_ar(p.min = 1, p.max = 8) \%>\%
 best_model(ic = "bic")
}
\author{
James Balamuta
}
