% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts.R
\name{comb}
\alias{comb}
\title{Combine math expressions}
\usage{
comb(...)
}
\arguments{
\item{...}{Expressions to combine.}
}
\value{
A combined expression.
}
\description{
Combine math expressions
}
\author{
Stephane Guerrier
}
\keyword{internal}
