% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{corr_analysis}
\alias{corr_analysis}
\title{Correlation Analysis Functions}
\usage{
corr_analysis(x, lagmax = 0, cor = TRUE, demean = TRUE, show.ci = TRUE,
  alpha = 0.05, plot = TRUE, ...)
}
\arguments{
\item{x}{An \code{"ts"} object.}

\item{lagmax}{A \code{integer} indicating the max lag.}

\item{cor}{A \code{bool} indicating whether the correlation (\code{TRUE}) or covariance (\code{FALSE}) should be computed.}

\item{demean}{A \code{bool} indicating whether the data should be detrended (\code{TRUE}) or not (\code{FALSE}).}

\item{show.ci}{A \code{bool} indicating whether to show confidence region.}

\item{alpha}{A \code{double} indicating the confidence interval level. Default is 0.05.}

\item{plot}{A \code{bool} indicating whether to plot.}

\item{...}{Additional parameters.}
}
\value{
Two \code{array} of dimensions \eqn{N \times S \times S}{N x S x S}.
}
\description{
Correlation Analysis function computes and plots both empirical ACF and PACF
or both univariate and multivariate cases.
}
\examples{
# Plot the Partial Autocorrelation
corr_analysis(datasets::AirPassengers)
}
\author{
Yunxiang Zhang
}
