% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portmanteau_tests.R
\name{diag_boxpierce}
\alias{diag_boxpierce}
\alias{diag_boxpierce.Arima}
\alias{diag_boxpierce.default}
\title{Box-Pierce}
\usage{
diag_boxpierce(x, order = NULL, stop_lag = 20, stdres = FALSE, ...)

\method{diag_boxpierce}{Arima}(x, stop_lag = 20, stdres = FALSE, ...)

\method{diag_boxpierce}{default}(x, order = NULL, stop_lag = 20,
  stdres = FALSE, ...)
}
\arguments{
\item{x}{An \code{arima} or data set.}

\item{order}{An \code{integer} indicating the degrees of freedom. If `x` is
not a series of residuals, then set equal to 0.}

\item{stop_lag}{An \code{integer} indicating the length of lags that should
be calculated.}

\item{stdres}{A \code{boolean} indicating whether to standardize the
residualizes (e.g. \eqn{res/sd(res)}) or not.}

\item{...}{Additional parameters}
}
\description{
Performs the Box-Pierce test to assess the Null Hypothesis of Independence
 in a Time Series
}
\author{
James Balamuta
}
