% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diag_plot}
\alias{diag_plot}
\title{Diagnostic Plot of Residuals}
\usage{
diag_plot(Xt, model, std = FALSE)
}
\arguments{
\item{Xt}{The original time series data.}

\item{model}{The \code{arima} model fit to the data.}

\item{std}{A \code{boolean} indicating whether we use standardized residuals for 
the (1) residuals plot and the (8) Ljung-Box test results.}
}
\description{
This function will plot eight diagnostic plots to assess how well the model fits  
the data. These include: (1) residuals plot, (2) residuals vs fitted values, 
(3) histogram of standardized residuals, (4) normal Q-Q plot of 
residuals, (5) ACF plot, (6) PACF plot, (7) Haar Wavelet variance representation,
(8) Ljung-Box test results.
}
\examples{
Xt = gen_gts(300, AR(phi = c(0, 0, 0.8), sigma2 = 1))
model = arima(Xt, order = c(3,0,0), include.mean = TRUE)
diag_plot(Xt, model)
}
\author{
Yuming Zhang
}
