% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portmanteau_tests.R
\name{plot.diag_portmanteau}
\alias{plot.diag_portmanteau}
\title{Graph Portmanteau Test Results}
\usage{
\method{plot}{diag_portmanteau}(object, ...)
}
\arguments{
\item{object}{A \code{diag_portmanteau} object from either
\code{\link{diag_boxpierce}} or \code{\link{diag_ljungbox}}}
}
\description{
Plots the portmanteau test results
}
\author{
Yuming Zhang
}
