% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsimts.R
\name{estimate}
\alias{estimate}
\title{Fit a Time Series Model to Data}
\usage{
estimate(model, Xt, method = "mle", demean = TRUE)
}
\arguments{
\item{model}{A time series model.}

\item{Xt}{A \code{vector} of time series data.}

\item{method}{A \code{string} indicating the method used for model fitting. 
Supported methods include \code{mle}, \code{yule-walker}, \code{gmwm}  and \code{rgmwm}.}

\item{demean}{A \code{boolean} indicating whether the model includes a mean / intercept term or not.}
}
\description{
This function can fit a time series model to data using different methods.
}
\examples{
Xt = gen_gts(300, AR(phi = c(0, 0, 0.8), sigma2 = 1))
model = estimate(AR(3), Xt)
model = estimate(AR(3), Xt, method = "rgmwm")
}
\author{
Stéphane Guerrier and Yuming Zhang
}
