% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{gen_fgn}
\alias{gen_fgn}
\title{Generate a Fractional Gaussian noise given \eqn{\sigma^2}{sigma^2} and \eqn{\H}{H}.}
\usage{
gen_fgn(N, sigma2 = 1, H = 0.9)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{sigma2}{A \code{double}.}

\item{H}{A \code{double}.}
}
\value{
fgn A \code{vec} containing the Fractional Gaussian noise process.
}
\description{
Simulates a Fractional Gaussian noise given \eqn{\sigma^2} and \eqn{\H}.
}
\keyword{internal}
