% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{gen_matern}
\alias{gen_matern}
\title{Generate a Matern Process given \eqn{\sigma^2}, \eqn{\lambda} and \eqn{\alpha}.}
\usage{
gen_matern(N, sigma2 = 1, lambda = 0.35, alpha = 0.9)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{sigma2}{A \code{double}.}

\item{lambda}{A \code{double}.}

\item{alpha}{A \code{double}.}
}
\value{
mtp A \code{vec} containing the Matern Process.
}
\description{
Simulates a Matern Process given \eqn{\sigma^2}, \eqn{\lambda} and \eqn{\alpha}.
}
\keyword{internal}
