% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{gen_mean}
\alias{gen_mean}
\title{Generate a determinist vector returned by the matrix by vector product of matrix \eqn{X} and vector \eqn{\beta}.}
\usage{
gen_mean(X, beta)
}
\arguments{
\item{X}{A \code{Matrix}  with dimension n*p.}

\item{beta}{A \code{vector} with dimension p*1}
}
\value{
mean_vec A \code{vec} containing the determinist vector.
}
\description{
Generate a determinist vector returned by the matrix by vector product of matrix \eqn{X} and vector \eqn{\beta}.
}
\keyword{internal}
