% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{gen_powerlaw}
\alias{gen_powerlaw}
\title{Generate a Power Law Process given \eqn{\sigma^2} and \eqn{\d}.}
\usage{
gen_powerlaw(N, sigma2 = 1, d = 0.9)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{sigma2}{A \code{double}.}

\item{d}{A \code{double}.}
}
\value{
plp A \code{vec} containing the Power Law Process.
}
\description{
Simulates a a Power Law Process given \eqn{\sigma^2} and \eqn{\d}.
}
\keyword{internal}
