% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts.R
\name{gts_time}
\alias{gts_time}
\title{Time of a gts object}
\usage{
gts_time(x)
}
\value{
Time vector of a gts object.
}
\description{
Extracting the time of a gts object
}
\author{
Stéphane Guerrier
}
\keyword{internal}
