% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_install.R
\name{install_smacdata}
\alias{install_smacdata}
\title{Install SMAC Data Package}
\usage{
install_smacdata()
}
\description{
Downloads and Install the SMAC Data Package to use with the simts Package
}
\details{
The SMAC Data Package contains data sets used in various packages.
}
\examples{
\dontrun{
# Online install
install_smacdata()
}
}
\author{
Justin Lee
}
