% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{M}
\alias{M}
\title{Definition of a Mean deterministic vector returned by the matrix by vector product of matrix \eqn{X} and vector \eqn{\beta}}
\usage{
M(X, beta)
}
\arguments{
\item{X}{A \code{Matrix}  with dimension n*p.}

\item{beta}{A \code{vector} with dimension p*1}
}
\value{
An S3 object containing the specified ts.model with the following structure:
\describe{
 \item{process.desc}{Used in summary: "X","BETA"}
 \item{theta}{Matrix X, vector beta}
 \item{plength}{Number of parameters}
 \item{print}{String containing simplified model}
 \item{desc}{"M"}
 \item{obj.desc}{Depth of Parameters e.g. list(1,1)}
 \item{starting}{Find starting values? TRUE or FALSE (e.g. specified value)}
}
}
\description{
Definition of a Mean deterministic vector returned by the matrix by vector product of matrix \eqn{X} and vector \eqn{\beta}
}
\examples{
X = matrix(rnorm(15*5), nrow = 15, ncol = 5)
beta=seq(5)
M(X = X, beta = beta)
}
\author{
Lionel Voirol,  Davide Cucci
}
