% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{Ma_cpp}
\alias{Ma_cpp}
\title{Ma function.}
\usage{
Ma_cpp(x, alpha)
}
\arguments{
\item{x}{A \code{double}.}

\item{alpha}{A \code{double}.}
}
\description{
Ma function.
}
\keyword{internal}
