% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{Ma_cpp_vec}
\alias{Ma_cpp_vec}
\title{Ma vectorized function.}
\usage{
Ma_cpp_vec(x, alpha)
}
\arguments{
\item{x}{A \code{NumericVector}.}
}
\description{
Ma vectorized function.
}
\keyword{internal}
