% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsimts.R
\name{MAPE}
\alias{MAPE}
\title{TO DO}
\usage{
MAPE(model, Xt, start = 0.25)
}
\arguments{
\item{model}{TO DO}

\item{Xt}{TO DO}

\item{start}{TO DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stéphane Guerrier and Yuming Zhang
}
