% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{MAT}
\alias{MAT}
\title{Definition of a Matérn Process}
\usage{
MAT(sigma2 = 1, lambda = 0.35, alpha = 0.9)
}
\arguments{
\item{sigma2}{A \code{double}.}

\item{lambda}{A \code{double}.}

\item{alpha}{A \code{double}.}
}
\value{
An S3 object containing the specified ts.model with the following structure:
\describe{
 \item{process.desc}{Used in summary: "SIGMA2","LAMBDA""ALPHA"}
 \item{theta}{Parameter vector including \eqn{\sigma^2}{sigma2},  \eqn{\lambda}{lambda},\eqn{\alpha}{alpha}}
 \item{plength}{Number of parameters}
 \item{print}{String containing simplified model}
 \item{desc}{"MAT"}
 \item{obj.desc}{Depth of Parameters e.g. list(1,1,1)}
 \item{starting}{Find starting values? TRUE or FALSE (e.g. specified value)}
}
}
\description{
Definition of a Matérn Process
}
\examples{
MAT()
MAT(sigma2 = 1, lambda = 0.35, alpha = 0.9)
}
\author{
Lionel Voirol,  Davide Cucci
}
