% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nile}
\alias{nile}
\title{nile}
\format{Time Series of length 100

For further details, see \url{https://stat.ethz.ch/R-manual/R-devel/library/datasets/html/Nile.html}}
\usage{
nile
}
\description{
Measurements of the annual flow of the river Nile at Aswan (formerly Assuan)
}
\keyword{datasets}
