% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsimts.R
\name{np_boot_sd_med}
\alias{np_boot_sd_med}
\title{Bootstrap standard error for the median}
\usage{
np_boot_sd_med(x, B = 5000)
}
\arguments{
\item{\code{x}}{A \code{vector} of data.}

\item{\code{B}}{A \code{numeric}. Number of simulations.}
}
\value{
Bootstrap standard error for the median
}
\description{
Non-parametric bootstrap to obtain the standard of the median of
iid data.
}
