% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{plot.ACF}
\alias{plot.ACF}
\title{Plot Auto-Covariance and Correlation Functions}
\usage{
\method{plot}{ACF}(x, xlab = NULL, ylab = NULL, show.ci = TRUE,
  alpha = NULL, col_ci = NULL, transparency = NULL, main = NULL,
  parValue = NULL, ...)
}
\arguments{
\item{x}{An \code{"ACF"} object from \code{\link{ACF}}.}

\item{xlab}{A \code{string} indicating the label of x axis, the default name is 'Lags'}

\item{ylab}{A \code{string} indicating the label of y axis, the default name is 'ACF'.}

\item{show.ci}{A \code{bool} indicating whether to show confidence region.}

\item{alpha}{A \code{double} indicating the confidence interval level. Default is 0.05.}

\item{col_ci}{A \code{string} that specifies the color of the confidence interval polygon.}

\item{transparency}{A \code{double} between 0 and 1 indicating the transparency level of the confidence region.
Default is 0.25.}

\item{main}{A \code{string} indicating the title of the plot. Default name is "Variable name - ACF plot'.}

\item{...}{Additional parameters}
}
\value{
An \code{array} of dimensions \eqn{N \times S \times S}{N x S x S}.
}
\description{
The acf function computes the estimated
autocovariance or autocorrelation for both univariate and multivariate cases.
}
\examples{
# Calculate the Autocorrelation
m = ACF(datasets::AirPassengers)

# Plot with 95\% CI
plot(m) 

# Plot with 90\% CI
plot(m, ci = 0.90) 

# Plot without 95\% CI
plot(m, show.ci = FALSE)

# More customized CI
plot(m, xlab = "my xlab", ylab = "my ylab", show.ci = TRUE,
alpha = NULL, col_ci = "grey", transparency = 0.5, main = "my main")
}
\author{
Yunxiang Zhang, Stéphane Guerrier and Yuming Zhang
}
