% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts.R
\name{plot.gts}
\alias{plot.gts}
\title{Plot Time Series Data}
\usage{
\method{plot}{gts}(object, xlab = "Time", ylab = "Observation",
  main = NULL, col = "dodgerblue", ...)
}
\arguments{
\item{object}{A \code{gts} object}

\item{xlab}{A \code{string} that gives a title for the x axis.}

\item{ylab}{A \code{string} that gives a title for the y axis.}

\item{main}{A \code{string} that gives an overall title for the plot.}

\item{col}{A \code{string} that gives a color for the line.}

\item{...}{other arguments passed to specific methods}
}
\value{
A plot containing the graph of the gts time series.
}
\description{
Plot Time Series Data generated by gts or gen_gts.
}
\author{
Justin Lee
}
\keyword{internal}
