% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_gen.R
\name{plot.KF}
\alias{plot.KF}
\title{Plot Kalman filter estimate}
\usage{
\method{plot}{KF}(x, plot_state = "all", ...)
}
\arguments{
\item{x}{A \code{KF} object resulting from the function \code{kalman_filter}}

\item{plot_state}{A numeric value indicating which state to plot. Default is "all".}

\item{...}{Additional arguments affecting the plot produced.}
}
\description{
The function plots the observed time serie, the sum of the estimated states and the confidence intervals of the estimate
}
\examples{
#Filter a 2*AR1 + DR + RW + WN process
set.seed(123)
model = AR(.3, 2) + AR(.5,3) + DR(.1) + RW(3) + WN(4) 
n = 250
y = gen_lts(n = n, model = model)
my_res = kalman_filter(model = model, y = y, h = 10)
plot(my_res)
plot(my_res, plot_state = 3)
}
\author{
Lionel Voirol
}
