% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts.R
\name{plot.lts}
\alias{plot.lts}
\title{Plot Latent Time Series Object}
\usage{
\method{plot}{lts}(x, xlab = NULL, ylab = NULL, main = NULL,
  color = NULL, fixed_range = FALSE, ...)
}
\arguments{
\item{x}{A \code{lts} object}

\item{xlab}{A \code{string} that gives a title for the x axis.}

\item{ylab}{A \code{string} that gives a title for the y axis.}

\item{main}{A \code{string} that gives an overall title for the plot.}

\item{color}{A \code{string} that gives a color for the line.}

\item{...}{additional arguments affecting the plot produced.}
}
\value{
A plot containing the graph of the latent time series.
}
\description{
Plot Latent Time Series Data included in an \code{lts} object.
}
\author{
Stephane Gurrier and Justin Lee
}
\keyword{internal}
