% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{plot.PACF}
\alias{plot.PACF}
\title{Plot Partial Auto-Covariance and Correlation Functions}
\usage{
\method{plot}{PACF}(object, show.ci = TRUE, alpha = 0.05, main = NULL,
  ...)
}
\arguments{
\item{show.ci}{A \code{bool} indicating whether to show confidence region}

\item{...}{Additional parameters}

\item{x, object}{An \code{"PACF"} object from \code{\link{PACF}}.}

\item{ci}{A \code{double} containing the 1-alpha level. Default is 0.95}
}
\value{
An \code{array} of dimensions \eqn{N \times S \times S}{N x S x S}.
}
\description{
Plot pacf function computes the estimated
plot partial autocovariance or autocorrelation for both univariate and multivariate cases.
}
\examples{
# Plot the Partial Autocorrelation
m = PACF(datasets::AirPassengers)
plot.PACF(m)
}
\author{
Yunxiang Zhang
}
