% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{PLP}
\alias{PLP}
\title{Definition of a Power Law Process}
\usage{
PLP(sigma2 = 1, d = 0.4)
}
\arguments{
\item{sigma2}{A \code{double}.}

\item{d}{A \code{double}.}
}
\value{
An S3 object containing the specified ts.model with the following structure:
\describe{
 \item{process.desc}{Used in summary: "SIGMA2","d"}
 \item{theta}{Parameter vector including \eqn{\sigma^2}{sigma2}, \eqn{\d}{H}}
 \item{plength}{Number of parameters}
 \item{print}{String containing simplified model}
 \item{desc}{"PLP"}
 \item{obj.desc}{Depth of Parameters e.g. list(1,1)}
 \item{starting}{Find starting values? TRUE or FALSE (e.g. specified value)}
}
}
\description{
Definition of a Power Law Process
}
\examples{
PLP()
PLP(sigma2 = 1, d = 0.4)
}
\author{
Lionel Voirol,  Davide Cucci
}
