% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsimts.R
\name{print.fitsimts}
\alias{print.fitsimts}
\title{Print fitsimts object}
\usage{
\method{print}{fitsimts}(out, ...)
}
\arguments{
\item{out}{A \code{fitsimts} object}
}
\value{
Text output via print
}
\description{
This function displays the information of a fitsimts object.
}
\author{
Stéphane Guerrier and Yuming Zhang
}
\keyword{internal}
