% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwm.r
\name{print.gmwm}
\alias{print.gmwm}
\title{Print gmwm object}
\usage{
\method{print}{gmwm}(x, ...)
}
\arguments{
\item{x}{A \code{GMWM} object}

\item{...}{Other arguments passed to specific methods}
}
\value{
Text output via print
}
\description{
Displays information about GMWM object
}
\author{
JJB
}
\keyword{internal}
