% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmwm.r
\name{print.summary.gmwm}
\alias{print.summary.gmwm}
\title{Print summary.gmwm object}
\usage{
\method{print}{summary.gmwm}(x, ...)
}
\arguments{
\item{x}{A \code{GMWM} object}

\item{...}{Other arguments passed to specific methods}
}
\value{
Text output via print
}
\description{
Displays summary information about GMWM object
}
\author{
JJB
}
\keyword{internal}
