% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{resid_plot}
\alias{resid_plot}
\title{Plot the Distribution of (Standardized) Residuals}
\usage{
resid_plot(Xt, model, std = FALSE, type = "hist", ...)
}
\arguments{
\item{Xt}{The data used to construct said model.}

\item{model}{The \code{arima} model used to fit the data.}

\item{std}{A \code{boolean} indicating whether the residuals plot is for standardized
residuals or original residuals.}

\item{type}{A \code{string} indicating either:
\code{"hist"} (standardized residual histogram), \code{"resid"} (residual plot),
or \code{"both"}}
}
\description{
This function can plot a histogram of the distribution of standardized
residuals, or plot the (standardized) residuals plot, or both.
}
\examples{
Xt = gen_gts(300, AR(phi = c(0, 0, 0.8), sigma2 = 1))
model = arima(Xt, order = c(3,0,0), include.mean = TRUE)

resid_plot(Xt, model, type = "hist")
resid_plot(Xt, model, type = "resid")
resid_plot(Xt, model, std = TRUE, type = "both")
}
\author{
Yuming Zhang
}
