% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RW2dimension.R
\name{RW2dimension}
\alias{RW2dimension}
\title{Function to Compute Direction Random Walk Moves}
\usage{
RW2dimension(steps = 100, probs = c(0.25, 0.5, 0.75))
}
\arguments{
\item{steps}{An \code{integer} that counts the number of steps of the random walk.}

\item{probs}{A \code{vector} of \code{double} that specifies the probabilities to choose each direction.}
}
\description{
The RW2dimension function computes direction random walk moves.
}
\examples{
RW2dimension(steps = 50, probs = c(0.2, 0.5, 0.6))
}
\author{
Stéphane Guerrier
}
