% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sales}
\alias{sales}
\title{Sales Dataset}
\format{A dataframe with 302 rows and 1 variable:}
\source{
\url{http://r-exercises.com/wp-content/uploads/2017/04/sales.csv}
}
\usage{
sales
}
\description{
This dataset contains the US monthly clothing retail sales in millions of dollars taken from 
1992 to 2016 for a total of 302 observations.
}
\keyword{datasets}
