% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{savingrt}
\alias{savingrt}
\title{savingrt}
\format{gts time series with frequency 12 starting in year 1959 and going
to 2016 for a total of 691 observations.
For further details, see \url{https://fred.stlouisfed.org/series/PSAVERT}}
\usage{
savingrt
}
\description{
Personal saving as a percentage of disposable personal
income (DPI), frequently referred to as "the personal saving rate,"
is calculated as the ratio of personal saving to DPI.
}
\keyword{datasets}
