% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{select_ar}
\alias{select_ar}
\title{Run Model Selection Criteria on AR Models}
\usage{
select_ar(xt, p.min = 1L, p.max = 5L, include.mean = TRUE)
}
\arguments{
\item{xt}{A data set}

\item{p.min}{Lowest Order AR(P) process to search}

\item{p.max}{Highest Order AR(P) process to search}

\item{include.mean}{A \code{boolean} indicating whether to fit AR with the mean or not}
}
\description{
Performs model fitting and calculates model selection criteria
}
\author{
James Balamuta and Stéphane Guerrier
}
