% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{select_arima}
\alias{select_arima}
\alias{select_arma}
\alias{select_ar}
\alias{select_ma}
\title{Run Model Selection Criteria on ARIMA Models}
\usage{
select_arima(xt, p.min = 0L, p.max = 5L, d.min = 0L, d.max = 2L,
  q.min = 0L, q.max = 5L, include.mean = TRUE)

select_arma(xt, p.min = 1L, p.max = 5L, q.min = 1L, q.max = 5L,
  include.mean = TRUE)

select_ar(xt, p.min = 1L, p.max = 5L, include.mean = TRUE)

select_ma(xt, q.min = 1L, q.max = 5L, include.mean = TRUE)
}
\arguments{
\item{xt}{A \code{vector} of univariate time series.}

\item{p.min}{An \code{integer} indicating the lowest order of AR(p) process to search.}

\item{p.max}{An \code{integer} indicating the highest order of AR(p) process to search.}

\item{d.min}{An \code{integer} indicating the lowest difference of data to take.}

\item{d.max}{An \code{integer} indicating the highest difference of data to take.}

\item{q.min}{An \code{integer} indicating the lowest order of MA(q) process to search.}

\item{q.max}{An \code{integer} indicating the highest order of MA(q) process to search.}

\item{include.mean}{A \code{bool} indicating whether to fit ARIMA with the mean or not.}
}
\description{
This function performs model fitting and calculates the model selection criteria.
}
\examples{
xt = gen_ar1(100, 0.3, 1)
x = select_arima(xt)
plot(x)

xt = gen_ma1(100, 0.3, 1)
x = select_ma(xt, q.min=2L, q.max=5L)
plot(x)

xt = gen_arma(10, c(.3,.5), c(.1), 1, 0)  
x = select_arma(xt, p.min = 1L, p.max = 4L,
                q.min = 1L, q.max = 3L)
plot(x)
}
