% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{select_ma}
\alias{select_ma}
\title{Run Model Selection Criteria on MA Models}
\usage{
select_ma(xt, q.min = 1, q.max = 10, include.mean = TRUE)
}
\arguments{
\item{xt}{A data set}

\item{q.min}{Lowest Order MA(Q) process to search}

\item{q.max}{Highest Order MA(Q) process to search}

\item{include.mean}{A \code{boolean} indicating whether to fit MA with the mean or not}
}
\description{
Performs model fitting and calculates model selection criteria
}
\author{
James Balamuta and Stéphane Guerrier
}
