% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsimts.R
\name{summary.fitsimts}
\alias{summary.fitsimts}
\title{Summary of fitsimts object}
\usage{
\method{summary}{fitsimts}(object, ...)
}
\arguments{
\item{object}{A \code{fitsimts} object}

\item{...}{Other arguments passed to specific methods}
}
\value{
Estimated parameters values with confidence intervals and standard errors.
}
\description{
Displays summary information about fitsimts object
}
\author{
Stéphane Guerrier
}
