% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{theo_acf}
\alias{theo_acf}
\title{Theoretical Autocorrelation (ACF) of an ARMA process}
\usage{
theo_acf(model = ARMA(ar = c(0.5, -0.25), ma = 0.2), lagmax = 20)
}
\arguments{
\item{lagmax}{A \code{integer} indicating the max length.}

\item{ar}{A \code{vector} containing the AR coefficients.}

\item{ma}{A \code{vector} containing the MA coefficients.}
}
\description{
This function computes the theoretical Autocorrelation (ACF) of an ARMA process.
}
\examples{
# Computes the theoretical ACF for an ARMA(1,0) or better known as an AR(1)
theo_acf(ARMA(ar = -0.25, ma = NULL))
# Computes the theoretical ACF for an ARMA(2, 1)
theo_acf(ARMA(ar = c(.50, -0.25), ma = 0.20), lagmax = 10)
}
\author{
Yuming Zhang
}
