% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{ACF}
\alias{ACF}
\title{Estimation of Auto-Covariance and Correlation Functions}
\usage{
ACF(x, lagmax = 0, cor = TRUE, demean = TRUE)
}
\arguments{
\item{x}{A \code{vector} or \code{ts} object (of length \eqn{N > 1}).}

\item{lagmax}{An \code{integer} indicating the maximum lag up to which to compute the empirical ACF.}

\item{cor}{A \code{bool} indicating whether the correlation (\code{TRUE}) or covariance (\code{FALSE}) should be computed. Defaults to \code{TRUE}.}

\item{demean}{A \code{bool} indicating whether the data should be detrended (\code{TRUE}) or not (\code{FALSE}). Defaults to \code{TRUE}.}
}
\value{
An \code{array} of dimensions \eqn{N \times 1 \times 1}{N x 1 x 1}.
}
\description{
The ACF function estimates the
autocovariance or autocorrelation for univariate time series.
}
\details{
\code{lagmax} default is \eqn{10*log10(N/m)} where \eqn{N} is the number of
observations and \eqn{m} is the number of time series being compared. If 
\code{lagmax} supplied is greater than the number of observations N, then one
less than the total will be taken (i.e. N - 1).
}
\examples{
# Get Autocorrelation
m = ACF(datasets::AirPassengers)

# Get Autocovariance and do not remove trend from signal
m = ACF(datasets::AirPassengers, cor = FALSE, demean = FALSE)
}
\author{
Yunxiang Zhang
}
