% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{best_model}
\alias{best_model}
\title{Select the Best Model}
\usage{
best_model(x, ic = "aic")
}
\arguments{
\item{x}{An object from either \code{\link{select_arima}},
\code{\link{select_arma}}, \code{\link{select_ar}}, or \code{\link{select_ma}}.}

\item{ic}{A \code{string} indicating the type of criterion to use in selecting the best model. 
Supported criteria include "AIC", "BIC" and "HQ".}
}
\description{
This function retrieves the best model.
}
