% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hydro}
\alias{hydro}
\title{Mean Monthly Precipitation, from 1907 to 1972}
\format{A time series object with frequency 12 starting at 1907 and going to 1972
for a total of 781 observations.}
\source{
\url{https://datamarket.com/data/set/22w1/mean-monthly-precipitation-1907-1972}
}
\usage{
hydro
}
\description{
Hydrology data that indicates a robust approach may be preferred to a
classical approach when estimating time series.
}
\keyword{datasets}
