% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_frame}
\alias{make_frame}
\title{Default utility function for various plots titles}
\usage{
make_frame(start, end, main, xlab, inloop = FALSE, i = 1)
}
\arguments{
\item{start}{A \code{numeric} that provides the time of the first observation.}

\item{end}{A \code{numeric} that provides the time of the last observation.}

\item{main}{A \code{string} that gives an overall title for the plot.}

\item{xlab}{A \code{string} that gives a title for the x axis.}

\item{inloop}{A \code{boolean} that checks if the object is plotting within a loop.}

\item{i}{A \code{numeric} that provides an index for when inloop is TRUE.}
}
\value{
Added title, grid, and axes.
}
\description{
Adds title, grid, and required x- and y-axes.
}
\author{
Stephane Guerrier and Justin Lee
}
