% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{PACF}
\alias{PACF}
\title{Partial Auto-Covariance and Correlation Functions}
\usage{
PACF(x, lag.max = NULL, type = "correlation", demean = TRUE)
}
\arguments{
\item{x}{A \code{vector} or \code{ts} object (of length \eqn{N > 1}).}

\item{lag.max}{An \code{integer} indicating the maximum lag up to which to compute the empirical PACF.}

\item{type}{A \code{character} string giving the type of acf to be computed. Allowed values are "correlation" (the default) and "covariance".}

\item{demean}{A \code{bool} indicating whether the data should be detrended (\code{TRUE}) or not (\code{FALSE}). Defaults to \code{TRUE}.}
}
\value{
An \code{array} of dimensions \eqn{N \times 1 \times 1}{N x 1 x 1}.
}
\description{
The PACF function estimates the
partial autocovariance or autocorrelation for both univariate time series.
}
\examples{
# Get Autocorrelation
m = PACF(datasets::AirPassengers)

# Get Autocovariance and do not remove trend from signal
m = PACF(datasets::AirPassengers, cor = FALSE, demean = FALSE)
}
\author{
Yunxiang Zhang
}
