% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts.R
\name{plot.gts}
\alias{plot.gts}
\title{Plot simts Time Series Data}
\usage{
\method{plot}{gts}(x, evenly = TRUE, xlab = NULL, ylab = NULL,
  main = NULL, couleur = "blue4", ...)
}
\arguments{
\item{x}{A \code{gts} object}

\item{evenly}{A \code{boolean} indicating whether the time series is evenly spaced or not.}

\item{xlab}{A \code{string} that gives a title for the x axis.}

\item{ylab}{A \code{string} that gives a title for the y axis.}

\item{main}{A \code{string} that gives an overall title for the plot.}

\item{couleur}{A \code{string} that gives a couleuror for the line.}

\item{...}{additional arguments affecting the plot produced.}
}
\value{
A plot containing the graph of the simts time series.
}
\description{
Plot simts Time Series Data generated by gts or gen_gts.
}
\author{
Justin Lee and Stéphane Guerrier
}
\keyword{internal}
