% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter_gen.R
\name{plot.KF}
\alias{plot.KF}
\title{Plot Kalman filter estimate}
\usage{
\method{plot}{KF}(obj)
}
\description{
The function plots the observed time serie, the sum of the estimated states and the confidence intervals of the estimate
}
