% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{plot.PACF}
\alias{plot.PACF}
\title{Plot Partial Auto-Covariance and Correlation Functions}
\usage{
\method{plot}{PACF}(x, xlab = NULL, ylab = NULL, show.ci = TRUE,
  alpha = NULL, col_ci = NULL, transparency = NULL, main = NULL,
  parValue = NULL, ...)
}
\arguments{
\item{x}{An \code{"PACF"} object from \code{\link{PACF}}.}

\item{xlab}{A \code{string} indicating the label of x axis, the default name is 'Lags'}

\item{ylab}{A \code{string} indicating the label of y axis, the default name is 'PACF'.}

\item{show.ci}{A \code{bool} indicating whether to show confidence region.}

\item{alpha}{A \code{double} indicating the confidence interval level. Default is 0.05.}

\item{col_ci}{A \code{string} indicating the color of the confidence interval polygon.}

\item{transparency}{A \code{double} between 0 and 1 indicating the transparency level of the confidence region.
Default is 0.25.}

\item{main}{A \code{string} indicating the title of the plot. Default name is "Variable name PACF plot'.}

\item{...}{Additional parameters.}
}
\value{
An \code{array} of dimensions \eqn{N \times S \times S}{N x S x S}.
}
\description{
Plot pacf function computes the estimated
plot partial autocovariance or autocorrelation for both univariate and multivariate cases.
}
\examples{
# Plot the Partial Autocorrelation
m = PACF(datasets::AirPassengers)
plot(m)

# More customized CI
plot(m, xlab = "my xlab", ylab = "my ylab", show.ci = TRUE, 
alpha = NULL, col_ci = "grey", transparency = 0.5, main = "my main")
}
\author{
Yunxiang Zhang and Yuming Zhang
}
