% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{plot.PACF}
\alias{plot.PACF}
\title{Plot Partial Auto-Covariance and Correlation Functions}
\usage{
\method{plot}{PACF}(x, show.ci = TRUE, ylab = "PACF", alpha = 0.05,
  main = NULL, ...)
}
\arguments{
\item{x}{An \code{"ACF"} object from \code{\link{ACF}}.}

\item{show.ci}{A \code{bool} indicating whether to show confidence region.}

\item{ylab}{A \code{string} indicating the y-axis label name.}

\item{alpha}{A \code{double} indicating the confidence interval level. Default is 0.05.}

\item{main}{A \code{string} indicating the title of the plot.}

\item{...}{Additional parameters.}
}
\value{
An \code{array} of dimensions \eqn{N \times S \times S}{N x S x S}.
}
\description{
Plot pacf function computes the estimated
plot partial autocovariance or autocorrelation for both univariate and multivariate cases.
}
\examples{
# Plot the Partial Autocorrelation
m = PACF(datasets::AirPassengers)
plot(m)
}
\author{
Yunxiang Zhang
}
