% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{plot.select_arima}
\alias{plot.select_arima}
\title{Visualization of Model Selection}
\usage{
\method{plot}{select_arima}(x, main = NULL)
}
\arguments{
\item{x}{An object that is either of type \code{\link{select_ar}}
or \code{\link{select_ma}}.}
}
\description{
This function visualize the model comparison based on different model selection criteria.
}
\examples{
xt = gen_ar1(100, 0.3, 1)
x = select_ar(xt)
plot(x)

xt = gen_ma1(100, 0.3, 1)
x = select_ma(xt, q.min=2L, q.max=5L)
plot(x)

}
\author{
Yuming Zhang
}
