% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_plot.R
\name{plot_pred}
\alias{plot_pred}
\title{Plot Time Series Forecast Function}
\usage{
plot_pred(x, model, n.ahead, time.pred, xlab = NULL, ylab = NULL,
  main = NULL, ...)
}
\arguments{
\item{x}{A \code{gts} object}

\item{model}{A \code{ts} model}

\item{n.ahead}{An \code{integer} indicating number of units of time ahead for which to make forecasts}

\item{time.pred}{A \code{vector} indicating the time points for which the forecasts are made (\code{lenght(vector)} must be equal to \code{n.ahead})}
}
\description{
This function plots the time series output from a forecast method with approximate 68% and 95% confidence intervals.
}
\examples{
x = gts(as.vector(lynx), start = 1821, end = 1934, freq = 1, 
unit_ts = bquote(paste(10^8," ",m^3)), name_ts = "Numbers", 
unit_time = "year", data_name = "Annual Numbers of Lynx Trappings")

model = arima(x, c(2,0,0))
n.ahead = 20
time.pred = 1935:1954
plot_pred(x, model, n.ahead, time.pred)
}
\author{
Yuming Zhang
}
