% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_plot.R
\name{plot_pred}
\alias{plot_pred}
\title{Plot Time Series Forecast Function}
\usage{
plot_pred(x, model, n.ahead, level = NULL, xlab = NULL, ylab = NULL,
  main = NULL, ...)
}
\arguments{
\item{x}{A \code{gts} object}

\item{model}{A \code{ts} model}

\item{n.ahead}{An \code{integer} indicating number of units of time ahead for which to make forecasts}

\item{level}{A \code{double} or \code{vector} indicating confidence level of prediction interval.
By default, it uses the levels of 0.50 and 0.95.}

\item{xlab}{A \code{string} for the title of x axis}

\item{ylab}{A \code{string} for the title of y axis}

\item{main}{A \code{string} for the over all title of the plot}

\item{...}{Additional parameters}
}
\description{
This function plots the time series output from a forecast method with approximate 68% and 95% confidence intervals.
}
\author{
Yuming Zhang
}
