% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsimts.R
\name{print.fitsimts}
\alias{print.fitsimts}
\title{Print fitsimts object}
\usage{
\method{print}{fitsimts}(x, ...)
}
\arguments{
\item{x}{A \code{fitsimts} object}

\item{...}{Other arguments passed to specific methods}
}
\value{
Text output via print
}
\description{
This function displays the information of a fitsimts object.
}
\author{
Stéphane Guerrier and Yuming Zhang
}
\keyword{internal}
