% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{resid_plot}
\alias{resid_plot}
\title{Plot the Distribution of (Standardized) Residuals}
\usage{
resid_plot(Xt, model, std = FALSE, type = "hist", ...)
}
\arguments{
\item{Xt}{The original time series data.}

\item{model}{The \code{arima} model fit to the data.}

\item{std}{A \code{boolean} indicating whether the residuals plot is for standardized
residuals or original residuals.}

\item{type}{A \code{string} indicating either:
\code{"hist"} (standardized residual histogram with superimposed kernel density estimator and normal distribution), \code{"resid"} (standard residual plot),
or \code{"both"}}

\item{...}{Additional parameters}
}
\description{
This function plots a histogram (with kernel density function and normal distribution) of the standardized
residuals or a basic plot the (standardized) residuals, or both.
}
\examples{
Xt = gen_gts(300, AR(phi = c(0, 0, 0.8), sigma2 = 1))
model = arima(Xt, order = c(3,0,0), include.mean = TRUE)

resid_plot(Xt, model, type = "hist")
resid_plot(Xt, model, type = "resid")
resid_plot(Xt, model, std = TRUE, type = "both")
}
\author{
Yuming Zhang
}
