% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{resid_plot}
\alias{resid_plot}
\title{Plot the Distribution of (Standardized) Residuals}
\usage{
resid_plot(res, std = FALSE, type = "hist", ...)
}
\arguments{
\item{res}{A \code{vector} of residuals.}

\item{std}{A \code{boolean} indicating whether the residuals plot is for standardized
residuals or original residuals.}

\item{type}{A \code{string} indicating either:
\code{"hist"} (standardized residual histogram with superimposed kernel density estimator and normal distribution), \code{"resid"} (standard residual plot),
or \code{"both"}}

\item{...}{Additional parameters}
}
\description{
This function plots a histogram (with kernel density function and normal distribution) of the standardized
residuals or a basic plot the (standardized) residuals, or both.
}
\author{
Yuming Zhang
}
