% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsimts.R
\name{select}
\alias{select}
\title{Time Series Model Selection}
\usage{
select(model, Xt, include.mean = TRUE, criterion = "aic")
}
\arguments{
\item{model}{A time series model.}

\item{Xt}{A \code{vector} of time series data.}

\item{include.mean}{A \code{boolean} indicating whether to fit ARIMA with the mean or not.}

\item{criterion}{A \code{string} indicating the type of criterion to use in selecting the best model. 
Supported criteria include "aic" (AIC), "bic" (BIC) and "hq" (HQ).}
}
\description{
This function performs model fitting and calculates the model selection criteria to be plotted.
}
\examples{
set.seed(463)
Xt = gen_gts(100, AR(phi = c(0.2, -0.5, 0.4), sigma2 = 1))
select(AR(5), Xt)
}
\author{
Stéphane Guerrier and Yuming Zhang
}
