% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{select_arima}
\alias{select_arima}
\alias{select_arma}
\alias{select_ar}
\alias{select_ma}
\title{Run Model Selection Criteria on ARIMA Models}
\usage{
select_arima(xt, p.min = 0L, p.max = 5L, d.min = 0L, d.max = 2L,
  q.min = 0L, q.max = 5L, include.mean = TRUE)

select_arma(xt, p.min = 0L, p.max = 5L, q.min = 0L, q.max = 5L,
  include.mean = TRUE)

select_ar(xt, p.min = 1L, p.max = 5L, include.mean = TRUE)

select_ma(xt, q.min = 1, q.max = 10, include.mean = TRUE)
}
\arguments{
\item{xt}{A data set}

\item{p.min}{Lowest Order AR(P) process to search}

\item{p.max}{Highest Order AR(P) process to search}

\item{d.min}{Lowest difference of data to take}

\item{d.max}{Highest difference of data to take}

\item{q.min}{Lowest Order MA(Q) process to search}

\item{q.max}{Highest Order MA(Q) process to search}

\item{include.mean}{A \code{boolean} indicating whether to fit ARIMA with the mean or not}
}
\description{
Performs model fitting and calculates model selection criteria
}
\author{
James Balamuta
}
