% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{simple_diag_plot}
\alias{simple_diag_plot}
\title{Simple Version of Diagnostic Plot of Residuals}
\usage{
simple_diag_plot(Xt, model, std = FALSE)
}
\arguments{
\item{Xt}{The data used to construct said model.}

\item{model}{The \code{arima} model used to fit the data.}

\item{std}{A \code{boolean} indicating whether we use standardized residuals for 
(1) residuals plot and (2) histogram of distribution of (standardized) residuals.}
}
\description{
This function will plot 4 diagnostic plots to assess the model used to 
fit the data. These include: (1) residuals plot, (2) histogram of distribution of 
(standardized) residuals, (3) Normal Q-Q plot of residuals and (4) Residuals vs Fitted plot.
}
\examples{
Xt = gen_gts(300, AR(phi = c(0, 0, 0.8), sigma2 = 1))
model = arima(Xt, order = c(3,0,0), include.mean = TRUE)
simple_diag_plot(Xt, model)
}
\author{
Yuming Zhang
}
