% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{simple_diag_plot}
\alias{simple_diag_plot}
\title{Basic Diagnostic Plot of Residuals}
\usage{
simple_diag_plot(Xt, model, std = FALSE)
}
\arguments{
\item{Xt}{The original time series data.}

\item{model}{The \code{arima} model fit to the data.}

\item{std}{A \code{boolean} indicating whether we use standardized residuals for the 
(1) residuals plot and the (2) histogram of (standardized) residuals.}
}
\description{
This function will plot four diagnostic plots to assess how well the model fits 
the data. These plots are: (1) residuals plot, (2) histogram of 
(standardized) residuals, (3) normal Q-Q plot of residuals and (4) residuals vs fitted values plot.
}
\author{
Yuming Zhang
}
