% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{simplified_print_SARIMA}
\alias{simplified_print_SARIMA}
\title{Simplify and print SARIMA model}
\usage{
simplified_print_SARIMA(p, i, q, P, si, Q, s)
}
\arguments{
\item{p}{An \code{integer} denoting the length of \code{ar}.}

\item{i}{An \code{integer} containing the number of differences to be done.}

\item{q}{An \code{integer} denoting the length of \code{ma}.}

\item{P}{An \code{integer} denoting the length of \code{sma}.}

\item{si}{An \code{integer} containing the number of seasonal differences to be done.}

\item{Q}{An \code{integer} denoting the length of \code{sar}.}

\item{s}{An \code{integer} indicating the seasonal value of the data.}
}
\value{
An S3 object with the following structure:
\describe{
 \item{print}{String containing simplified model}
 \item{simplified}{Type of model (after simplification)}
}
}
\description{
Simplify and print SARIMA model
}
\author{
Stephane Guerrier
}
