% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simts_package.R
\docType{package}
\name{simts-package}
\alias{simts-package}
\title{Simulated Time Series (simts) Package}
\description{
Simulated Time Series (simts) Package
}
\details{
\tabular{ll}{
Package: \tab simts\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-05-28\cr
License: \tab GPL\cr
}
}
\author{
James Balamuta \email{balamut2@illinois.edu},
Stephane Guerrier \email{stephane@illinois.edu},
Roberto Molinari \email{roberto.molinari@unige.ch},
Wenchao Yang \email{wyang40@illinois.edu}
Justin Lee \email{munsheet93@gmail.com}
}
\keyword{internal}
