% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/simts_package.R
\docType{package}
\name{simts-package}
\alias{simts}
\alias{simts-package}
\alias{simts-package}
\title{simts: Simulate Various Time Series}
\description{
Generates various time series objects for use in other packages.

Simulated Time Series (simts) Package
}
\details{
The data sets in this package may change at a moments notice.

\tabular{ll}{
Package: \tab simts\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-05-28\cr
License: \tab GPL\cr
}
}
\author{
\strong{Maintainer}: Justin Lee \email{munsheet93@gmail.com}

Authors:
\itemize{
  \item Stephane Guerrier
  \item James Balamuta
  \item Wenchao Yang
}

Other contributors:
\itemize{
  \item Nathanael Claussen [contributor]
}


James Balamuta \email{balamut2@illinois.edu},
Stephane Guerrier \email{stephane@illinois.edu},
Roberto Molinari \email{roberto.molinari@unige.ch},
Wenchao Yang \email{wyang40@illinois.edu}
Justin Lee \email{munsheet93@gmail.com}
}
\keyword{internal}
