% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/simts_package.R
\docType{package}
\name{simts-package}
\alias{simts}
\alias{simts-package}
\title{simts: Time Series Simulation}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Generates various time series objects for use in other packages.
}
\details{
The data sets in this package may change at a moments notice.

\tabular{ll}{
Package: \tab simts\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-09-15\cr
License: \tab GPL\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SMAC-Group/simts}
  \item Report bugs at \url{https://github.com/SMAC-Group/simts/issues}
}

}
\author{
\strong{Maintainer}: Stephane Guerrier \email{stef.guerrier@gmail.com} [copyright holder]

Authors:
\itemize{
  \item James Balamuta
  \item Roberto Molinari
  \item Justin Lee
  \item Yuming Zhang
}

Other contributors:
\itemize{
  \item Wenchao Yang [contributor]
  \item Nathanael Claussen [contributor]
  \item Yunxiang Zhang [contributor]
}


James Balamuta \email{balamut2@illinois.edu},
Stephane Guerrier \email{stephane@psu.edu},
Roberto Molinari \email{roberto.molinari@hotmail.it},
Wenchao Yang \email{wyang40@illinois.edu}
Justin Lee \email{munsheet93@gmail.com}
Yuming Zhang \email{yfz5097@psu.edu}
}
\keyword{internal}
