% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/simts_package.R
\docType{package}
\name{simts-package}
\alias{simts}
\alias{simts-package}
\title{simts: Time Series Analysis Tools}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

A system contains easy-to-use tools as a support for time series analysis courses. In particular, it incorporates a technique called Generalized Method of Wavelet Moments (GMWM) as well as its robust implementation for fast and robust parameter estimation of time series models which is described, for example, in Guerrier et al. (2013) <doi: 10.1080/01621459.2013.799920>. More details can also be found in the paper linked to via the URL below.
}
\details{
The data sets in this package may change at a moments notice.

\tabular{ll}{
Package: \tab simts\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-09-15\cr
License: \tab GPL\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SMAC-Group/simts}
  \item \url{https://arxiv.org/pdf/1607.04543.pdf}
  \item Report bugs at \url{https://github.com/SMAC-Group/simts/issues}
}

}
\author{
\strong{Maintainer}: Stéphane Guerrier \email{stef.guerrier@gmail.com} [copyright holder]

Authors:
\itemize{
  \item James Balamuta [copyright holder]
  \item Roberto Molinari [copyright holder]
  \item Justin Lee
  \item Yuming Zhang
}

Other contributors:
\itemize{
  \item Wenchao Yang [contributor]
  \item Nathanael Claussen [contributor]
  \item Yunxiang Zhang [contributor]
  \item Christian Gunning [copyright holder]
  \item Romain Francois [copyright holder]
  \item Ross Ihaka [copyright holder]
  \item R Core Team [copyright holder]
}


James Balamuta \email{balamut2@illinois.edu},
Stephane Guerrier \email{stephane@psu.edu},
Roberto Molinari \email{roberto.molinari@hotmail.it},
Wenchao Yang \email{wyang40@illinois.edu}
Justin Lee \email{munsheet93@gmail.com}
Yuming Zhang \email{yfz5097@psu.edu}
}
\keyword{internal}
